MaterialAlertDialog

+++ app/src/main/java/com/example/stores/MainActivity.kt
@@ -1,10 +1,12 @@
 package com.example.stores
 
+import android.content.DialogInterface
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.Toast
 import androidx.recyclerview.widget.GridLayoutManager
 import com.example.stores.databinding.ActivityMainBinding
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import org.jetbrains.anko.doAsync
 import org.jetbrains.anko.uiThread
 
@@ -108,12 +110,21 @@ class MainActivity : AppCompatActivity(), OnClickListener, MainAux{
     }
 
     override fun onDeleteStore(storeEntity: StoreEntity) {
-        doAsync {
-            StoreAplication.dataBase.storeDao().deleteStore(storeEntity)
-            uiThread {
-                mAdapter.delete(storeEntity)
-            }
-        }
+        MaterialAlertDialogBuilder(this)
+            .setTitle(R.string.dialog_delete_title)
+            .setPositiveButton(R.string.dialog_delete_confirm, DialogInterface.OnClickListener { dialogInterface, i ->
+
+                doAsync {
+                    StoreAplication.dataBase.storeDao().deleteStore(storeEntity)
+                    uiThread {
+                        mAdapter.delete(storeEntity)
+                    }
+                }
+            })
+            .setNegativeButton(R.string.dialog_cancel, null)
+            .show()
+
+
     }
